⚡️ : move interface to model
