+++ foody_client_nextjs/domain/models/recipeModel.ts
@@ -1,0 +1,16 @@
+export interface RecipeWithIngredient {
+  id: string;
+  title: string;
+  thumbnail: string;
+  link: string;
+  userId: string;
+  authorId: string;
+  createdAt: string;
+  updatedAt: string;
+  ingredients: {
+    ingredient: {
+      id: string;
+      name: string;
+    };
+  }[];
+}
