+++ foody_client_nextjs/domain/repositories/AuthRepository.ts
@@ -1,4 +1,4 @@
-import { SignInResponse } from '../models/AuthModel';
+import { SignInResponse } from '../models/authModel';
 import { post } from '@/utils/fetch';
 export class AuthRepository {
   async signIn(idToken: string) {
