+++ foody_client_nextjs/views/RecipeList/useRecipeList.ts
@@ -2,13 +2,12 @@
 
 import { useEffect, useState } from 'react';
 
-import useRecipesQuery, {
-  RecipeWithIngredient,
-} from '@/hooks/queries/useRecipesWithIngredientQuery';
+import { RecipeWithIngredient } from '@/domain/models/recipeModel';
+import useRecipesWithIngredientQuery from '@/hooks/queries/useRecipesWithIngredientQuery';
 
 export default function useRecipeList() {
   const [recipes, setRecipes] = useState<RecipeWithIngredient[]>([]);
-  const { data, isLoading, error } = useRecipesQuery();
+  const { data, isLoading, error } = useRecipesWithIngredientQuery();
 
   useEffect(() => {
     if (data) {
